AllListingsFragment is calling to the ListingModel to get listings from the ServiceClient

+++ app/src/main/java/com/example/campusplate_android/Model/Types/Listing.java
@@ -1,14 +1,14 @@
 package com.example.campusplate_android.Model.Types;
 
 public class Listing {
-    public int listingId;
-    public int userId;
+    public double listingId;
+    public double userId;
     public String title;
     public String locationDescription;
     public double lat;
     public double lng;
-    public int creationTime;
-    public int quantity;
+    public double creationTime;
+    public double quantity;
     //TODO: Add image
 
     public Listing() {
@@ -22,7 +22,7 @@ public class Listing {
         this.quantity = 0;
     }
 
-    public Listing(int listingId, int userId, String title, String locationDescription, double lat, double lng, int creationTime, int quantity) {
+    public Listing(double listingId, double userId, String title, String locationDescription, double lat, double lng, double creationTime, double quantity) {
         this.listingId = listingId;
         this.userId = userId;
         this.title = title;
